+++ packages/components/src/private-apis.ts
@@ -8,6 +8,7 @@ import Theme from './theme';
 import { Tabs } from './tabs';
 import { kebabCase } from './utils/strings';
 import { lock } from './lock-unlock';
+import Badge from './badge';
 
 export const privateApis = {};
 lock( privateApis, {
@@ -17,4 +18,5 @@ lock( privateApis, {
 	Theme,
 	Menu,
 	kebabCase,
+	Badge,
 } );