+++ packages/icons/src/icon/stories/keywords.ts
@@ -1,7 +1,9 @@
 const keywords: Partial< Record< keyof typeof import('../../'), string[] > > = {
 	cancelCircleFilled: [ 'close' ],
-	cautionFilled: [ 'alert', 'caution', 'warning' ],
+	caution: [ 'alert', 'warning' ],
+	cautionFilled: [ 'alert', 'warning' ],
 	create: [ 'add' ],
+	error: [ 'alert', 'caution', 'warning' ],
 	file: [ 'folder' ],
 	seen: [ 'show' ],
 	thumbsDown: [ 'dislike' ],