+++ packages/components/CHANGELOG.md
@@ -16,6 +16,10 @@
 
 -   `BoxControl`: Better respect for the `min` prop in the Range Slider ([#67819](https://github.com/WordPress/gutenberg/pull/67819)).
 
+### Experimental
+
+-   Add new `Badge` component ([#66555](https://github.com/WordPress/gutenberg/pull/66555)).
+
 ## 29.0.0 (2024-12-11)
 
 ### Breaking Changes